Latest neural network version update.
Update neural network architecture and features used for prediction. Inputs consists of 70-day timeseries windows and outputs of14-day timeseries windows. The neural network architecture consists of a single Bidirectional-LSTM layer with 64 hidden units, connected to a dense layer with 32 hidden units, SELU activation function and Lecun-normal kernel initializer. A dropout layer with a value of 0.4 is used before the output dense layer.

After many hours of parameter tuning and optimization, the best results on average were achieved using SGD as optimizer and a mean squared error loss function. To avoid overfitting early stopping is used with a patience window of seven steps, monitoring the validation loss. With the current architecture and training parameters, an epoch size of 40 to 80 epochs appear to be appropriate.